pyproject.toml
[project]
name = "kit-collector"
version = "0.1.0"
description = ""
authors = [
    {name = "IKoposhilov",email = "ikoposhilov@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=2.3.0,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pydantic (>=2.12.3,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
